NativeMethods: setup shutdown handler only when init succeeds
Only set up the object with the `git_libgit2_shutdown` finalizer when
`git_libgit2_init` has succeeded.  This ensures that setting up the
finalizer is the last thing that we do in the static constructor for
`NativeMethods`, meaning that any exception trying to p/invoke
`git_libgit2_init` remains catch-able.